Create a new user account on a Windows machine.
This playbook assumes you have the necessary permissions to create a new user account, and it uses the Administrator account to execute the task. Replace the following variables with your desired values:

windows_servers: Replace this with the actual hostname or IP address of your Windows machine. You can define it in your inventory file.
new_username: The username you want to create.
new_user_password: The password for the new user.
new_user_groups: The groups you want to add the user to. This can be a list of groups.
Save this playbook in a file, for example, create_user.yml. 
To execute the playbook, use the following command:
ansible-playbook -i inventory.ini create_user.yml

Make sure to replace inventory.ini with the path to your Ansible inventory file.

Note: Ensure that you have WinRM properly configured on your Windows machine and Ansible installed on the control machine. The control machine should also have Python installed, and the Windows machine should allow remote connections.